requirements.optional.txt
[sbml]
python_libsbml

[containers]
docker >= 4.4

[logging]
capturer
